requirements.txt
streamlit==1.35.0
faiss-cpu==1.7.4
sentence-transformers==2.5.1
transformers==4.41.2
torch==2.0.1
python-docx==0.8.11
pdfplumber==0.6.2
python-dotenv>=1.0.0
langchain>=0.1.14
accelerate>=0.21.0